requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
ecdsa==0.18.0
falcon==3.1.1
falcon-auth2==0.1.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.21
typing-extensions==4.8.0
urllib3==2.0.5
